+++ Code snippets/src/main/java/app/HomeUseCaseFactory.java
@@ -24,10 +24,10 @@ import main.java.interface_adapter.home.HomeViewModel;
 import main.java.interface_adapter.notes.NotesViewModel;
 import main.java.interface_adapter.notes.OpenNotesController;
 import main.java.interface_adapter.notes.OpenNotesPresenter;
-import main.java.use_case.notes.OpenNotesInputBoundary;
-//import main.java.use_case.notes.OpenNotesInteractor;
-import main.java.use_case.notes.OpenNotesInteractor;
-import main.java.use_case.notes.OpenNotesOutputBoundary;
+import main.java.use_case.notes.open_notes.OpenNotesInputBoundary;
+//import main.java.use_case.notes.open_notes.OpenNotesInteractor;
+import main.java.use_case.notes.open_notes.OpenNotesInteractor;
+import main.java.use_case.notes.open_notes.OpenNotesOutputBoundary;
 import main.java.use_case.reminder.ReminderInputBoundary;
 import main.java.use_case.reminder.ReminderInteractor;
 import main.java.use_case.reminder.ReminderOutputBoundary;
