+++ Code snippets/src/main/java/interface_adapter/notes/OpenNotesPresenter.java
@@ -1,8 +1,8 @@
 package main.java.interface_adapter.notes;
 
 import main.java.interface_adapter.ViewManagerModel;
-import main.java.use_case.notes.OpenNotesOutputBoundary;
-import main.java.use_case.notes.OpenNotesOutputData;
+import main.java.use_case.notes.open_notes.OpenNotesOutputBoundary;
+import main.java.use_case.notes.open_notes.OpenNotesOutputData;
 
 public class OpenNotesPresenter implements OpenNotesOutputBoundary {
     private ViewManagerModel viewManagerModel;
