+++ Code snippets/src/main/java/use_case/courses/add_course/AddCourseInputBoundary.java
@@ -1,4 +1,4 @@
-package main.java.use_case.notes;
+package main.java.use_case.courses.add_course;
 
 public interface AddCourseInputBoundary {
     public void execute(AddCourseInputData addCourseInputData);
