+++ Code snippets/src/main/java/use_case/courses/add_course/AddCourseOutputData.java
@@ -1,4 +1,4 @@
-package main.java.use_case.notes;
+package main.java.use_case.courses.add_course;
 
 import main.java.entity.Student;
