+++ Code snippets/src/main/java/use_case/notes/create_notes/CreateNotesOutputBoundary.java
@@ -1,4 +1,4 @@
-package main.java.use_case.notes;
+package main.java.use_case.notes.create_notes;
 
 public interface CreateNotesOutputBoundary {
     void prepareSuccessView(CreateNotesOutputData notes);
