+++ Code snippets/src/main/java/use_case/notes/open_notes/OpenNotesInputBoundary.java
@@ -1,4 +1,4 @@
-package main.java.use_case.notes;
+package main.java.use_case.notes.open_notes;
 
 public interface OpenNotesInputBoundary {
     public void execute();
